He borrado la carpeta "lost" (me daba error al compilar)
