+++ src/Controlador/Main.java
@@ -1,0 +1,18 @@
+package Controlador;
+import Modelo.*;
+import Vista.*;
+
+import java.util.Scanner;
+
+
+public class Main {
+    public static void main(String[] args) {
+Scanner input= new Scanner(System.in);
+
+
+        System.out.println("Hello world!");
+
+
+
+    }
+}
