+++ src/Modelo/Datos.java
@@ -1,0 +1,15 @@
+package Modelo;
+
+/** Clase que contiene todos las operaciones para el controlador */
+
+
+public class Datos {
+
+
+
+
+
+
+
+
+}
