+++ src/Modelo/ListaArticulo.java
@@ -1,0 +1,21 @@
+package Modelo;
+
+import java.util.ArrayList;
+
+public class ListaArticulo extends Lista {
+    /**
+     * Metodo constructor
+     *
+     * @param lista
+     */
+    public ListaArticulo(ArrayList<Articulos> lista) {
+        super(lista);
+    }
+
+    @Override
+    public String toString() {
+        return "ListaArticulo{" +
+                "lista=" + lista +
+                '}';
+    }
+}
