+++ src/Modelo/ListaPedidos.java
@@ -1,0 +1,21 @@
+package Modelo;
+
+import java.util.ArrayList;
+
+public class ListaPedidos extends Lista{
+    /**
+     * Metodo constructor
+     *
+     * @param lista
+     */
+    public ListaPedidos(ArrayList<Pedidos> lista) {
+        super(lista);
+    }
+
+    @Override
+    public String toString() {
+        return "ListaPedidos{" +
+                "lista=" + lista +
+                '}';
+    }
+}
