+++ src/Vista/AñadirArticulo.java
@@ -1,0 +1,4 @@
+package Vista;
+
+public class AñadirArticulo {
+}
