+++ src/Vista/AñadirCliente.java
@@ -1,0 +1,4 @@
+package Vista;
+
+public class AñadirCliente {
+}
