+++ src/Vista/EliminarPedido.java
@@ -1,0 +1,4 @@
+package Vista;
+
+public class EliminarPedido {
+}
