+++ src/Vista/Menu.java
@@ -1,0 +1,4 @@
+package Vista;
+
+public class Menu {
+}
