+++ src/Vista/MostrarCliente.java
@@ -1,0 +1,4 @@
+package Vista;
+
+public class MostrarCliente {
+}
