--- src/lost/controlador/Control_Services.java
@@ -1,8 +1,0 @@
-package lost.controlador;
-
-import Modelo.*;
-import Vista.*;
-
-
-public class Control_Services {
-}
