--- src/lost/controlador/Main.java
@@ -1,18 +1,0 @@
-package lost.controlador;
-import Modelo.*;
-import Vista.*;
-
-import java.util.Scanner;
-
-
-public class Main {
-    public static void main(String[] args) {
-Scanner input= new Scanner(System.in);
-
-
-        System.out.println("Hello world!");
-
-
-
-    }
-}
