--- src/lost/modelo/Datos.java
@@ -1,15 +1,0 @@
-package lost.modelo;
-
-/** Clase que contiene todos las operaciones para el controlador */
-
-
-public class Datos {
-
-
-
-
-
-
-
-
-}
