--- src/lost/vista/AñadirCliente.java
@@ -1,4 +1,0 @@
-package lost.vista;
-
-public class AñadirCliente {
-}
