--- src/lost/vista/MostrarArticulo.java
@@ -1,4 +1,0 @@
-package lost.vista;
-
-public class MostrarArticulo {
-}
