--- src/lost/vista/MostrarCliente.java
@@ -1,4 +1,0 @@
-package lost.vista;
-
-public class MostrarCliente {
-}
